temp/robot/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
crc8==0.1.0
dataclasses-json==0.4.2
entrypoints==0.3
flake8==3.7.9
idna==2.9
injector==0.18.3
marshmallow==3.5.1
marshmallow-enum==1.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
nose2==0.9.2
numpy==1.18.3
opencv-contrib-python==3.4.8.29
pycodestyle==2.5.0
pyflakes==2.1.1
pyserial==3.4
python-engineio==3.12.1
python-socketio==4.5.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
stringcase==1.2.0
typing-extensions==3.7.4.2
typing-inspect==0.5.0
urllib3==1.25.9
websocket-client==0.57.0
